test(http): Extract auth tests from `ApplicationTest` to `ApplicationAuthTest` class.
