+++ tests/http/stateless/ApplicationTest.php
@@ -16,7 +16,6 @@ use yii2\extensions\psrbridge\tests\support\FactoryHelper;
 use yii2\extensions\psrbridge\tests\TestCase;
 
 use function array_filter;
-use function base64_encode;
 use function explode;
 use function ini_get;
 use function ini_set;
@@ -532,98 +531,6 @@ final class ApplicationTest extends TestCase
     /**
      * @throws InvalidConfigException if the configuration is invalid or incomplete.
      */
-    public function testReturnCredentialsWhenValidBasicAuthorizationHeaderIsPresent(): void
-    {
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => 'Basic ' . base64_encode('user:pass'),
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-
-        $responseData = Json::decode($response->getBody()->getContents());
-
-        self::assertIsArray(
-            $responseData,
-            "Response 'body' should be an array after decoding JSON response from 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-        self::assertSame(
-            'user',
-            $responseData['username'] ?? '',
-            "Should extract 'user' with correct 'Basic ' ('6' chars) prefix",
-        );
-        self::assertSame(
-            'pass',
-            $responseData['password'] ?? '',
-            "Should extract 'pass' with correct 'Basic ' ('6' chars) prefix",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnCredentialsWithMultibyteCharacters(): void
-    {
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => "basic\xC2\xA0" . base64_encode('user:pass'),
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-
-        $responseData = Json::decode($response->getBody()->getContents());
-
-        self::assertIsArray(
-            $responseData,
-            "Response 'body' should be an array after decoding JSON response from 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-        self::assertSame(
-            'user',
-            $responseData['username'] ?? '',
-            "Should handle multibyte characters in 'username'",
-        );
-        self::assertSame(
-            'pass',
-            $responseData['password'] ?? '',
-            "Should handle multibyte characters in 'password'",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
     public function testReturnHtmlErrorResponseWhenErrorHandlerActionIsInvalid(): void
     {
         $_SERVER = [
@@ -665,67 +572,6 @@ final class ApplicationTest extends TestCase
             self::normalizeLineEndings($response->getBody()->getContents()),
             "Response 'body' should contain error message about 'An Error occurred while handling another error' and " .
             "the InvalidRouteException when errorHandler action is invalid in 'StatelessApplication'.",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnJsonResponseWithCredentialsForSiteAuthRoute(): void
-    {
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => 'Basic ' . base64_encode('admin:admin'),
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            <<<JSON
-            {"username":"admin","password":"admin"}
-            JSON,
-            $response->getBody()->getContents(),
-            "Response 'body' should match expected JSON string '{\"username\":\"admin\",\"password\":\"admin\"}' " .
-            "for 'site/auth' route in 'StatelessApplication'.",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnJsonResponseWithNullCredentialsForMalformedAuthorizationHeader(): void
-    {
-        $_SERVER = [
-            'HTTP_authorization' => 'Basic foo:bar',
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route with malformed authorization header in " .
-            "'StatelessApplication'.",
-        );
-        self::assertSame(
-            <<<JSON
-            {"username":null,"password":null}
-            JSON,
-            $response->getBody()->getContents(),
-            "Response 'body' should match expected JSON string '{\"username\":null,\"password\":null}' for malformed " .
-            "authorization header on 'site/auth' route in 'StatelessApplication'.",
         );
     }
 
@@ -871,172 +717,6 @@ final class ApplicationTest extends TestCase
             'site/update/123',
             $request->getUri()->getPath(),
             "Request 'path' should be 'site/update/123' for 'site/update/123' route in 'StatelessApplication'.",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnNullCredentialsWhenAuthorizationHeaderHasInvalidBasicPrefix(): void
-    {
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => 'basix ' . base64_encode('user:pass'),
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route with 'basix' 'HTTP_AUTHORIZATION' header " .
-            "in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route with 'basix' " .
-            "'HTTP_AUTHORIZATION' header in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            <<<JSON
-            {"username":null,"password":null}
-            JSON,
-            $response->getBody()->getContents(),
-            "Response 'body' should return 'null' credentials when the 'HTTP_AUTHORIZATION' header does not start " .
-            "with 'Basic ' (case-insensitive) for 'site/auth' route in 'StatelessApplication'.",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnNullCredentialsWhenBasicAuthorizationHeaderHasInvalidBase64DueToMissingSpace(): void
-    {
-        $criticalBase64 = base64_encode('a:b'); // "YTpi"
-
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => "Basic{$criticalBase64}", // "BasicYTpi"
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-
-        $responseData = Json::decode($response->getBody()->getContents());
-
-        self::assertIsArray(
-            $responseData,
-            "Response 'body' should be an array after decoding JSON response from 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-        self::assertNull(
-            $responseData['username'] ?? null,
-            'Should be null when cutting at position 6 produces invalid base64',
-        );
-        self::assertNull(
-            $responseData['password'] ?? null,
-            'Should be null when cutting at position 6 produces invalid base64',
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnNullCredentialsWhenBasicAuthorizationHeaderLacksSpace(): void
-    {
-        $base64Token = base64_encode('user:pass'); // 'dXNlcjpwYXNz'
-
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => "Basic{$base64Token}", // 'BasicdXNlcjpwYXNz'
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-
-        $responseData = Json::decode($response->getBody()->getContents());
-
-        self::assertIsArray(
-            $responseData,
-            "Response 'body' should be an array after decoding JSON response from 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-        self::assertNull(
-            $responseData['username'] ?? null,
-            "Should be 'null' when 'Basic' lacks the required space after the scheme",
-        );
-        self::assertNull(
-            $responseData['password'] ?? null,
-            'Should be null when the Authorization header lacks a space after Basic (invalid base64 segment)',
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnPartialCredentialsWhenOnlyUsernameIsPresent(): void
-    {
-        $_SERVER = [
-            'PHP_AUTH_USER' => 'admin',
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route with only 'PHP_AUTH_USER' set in " .
-            "'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route with only " .
-            "'PHP_AUTH_USER' set in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            <<<JSON
-            {"username":"admin","password":null}
-            JSON,
-            $response->getBody()->getContents(),
-            "Response 'body' should return username 'admin' and 'null' password when only 'PHP_AUTH_USER' is " .
-            "present, confirming OR logic works correctly for 'site/auth' route in 'StatelessApplication'.",
         );
     }
 
@@ -1210,51 +890,6 @@ final class ApplicationTest extends TestCase
             201,
             $response->getStatusCode(),
             "Response 'status code' should be '201' for 'site/statuscode' route in 'StatelessApplication'.",
-        );
-    }
-
-    /**
-     * @throws InvalidConfigException if the configuration is invalid or incomplete.
-     */
-    public function testReturnUsernameOnlyWhenNoColonSeparatorInCredentials(): void
-    {
-        $_SERVER = [
-            'HTTP_AUTHORIZATION' => 'Basic ' . base64_encode('usernameonly'),
-            'REQUEST_METHOD' => 'GET',
-            'REQUEST_URI' => 'site/auth',
-        ];
-
-        $app = $this->statelessApplication();
-
-        $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
-
-        self::assertSame(
-            200,
-            $response->getStatusCode(),
-            "Response 'status code' should be '200' for 'site/auth' route in 'StatelessApplication'.",
-        );
-        self::assertSame(
-            'application/json; charset=UTF-8',
-            $response->getHeaderLine('Content-Type'),
-            "Response 'Content-Type' should be 'application/json; charset=UTF-8' for 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-
-        $responseData = Json::decode($response->getBody()->getContents());
-
-        self::assertIsArray(
-            $responseData,
-            "Response 'body' should be an array after decoding JSON response from 'site/auth' route in " .
-            "'StatelessApplication'.",
-        );
-        self::assertSame(
-            'usernameonly',
-            $responseData['username'] ?? '',
-            "Should return 'username' when no colon separator is present in credentials.",
-        );
-        self::assertNull(
-            $responseData['password'] ?? null,
-            "Should return 'null' 'password' when no colon separator is present in credentials.",
         );
     }
